coordinator: move meshapi server to authority
